move test function to proxy_test.go
