+++ wasm/proxy.go
@@ -2,14 +2,11 @@ package wasm
 
 import (
 	"fmt"
-	"testing"
 
 	"github.com/0chain/gosdk/core/version"
 
-	"github.com/0chain/gosdk/bls"
 	// "github.com/0chain/gosdk/miracl"
 	// "github.com/0chain/gosdk/core/encryption"
-	"github.com/0chain/gosdk/core/zcncrypto"
 
 	"syscall/js"
 
@@ -43,37 +40,6 @@ import (
 	/// delete.go imports
 	// All already imported or not needed.
 )
-
-var verifyPublickey = `041eeb1b4eb9b2456799d8e2a566877e83bc5d76ff38b964bd4b7796f6a6ccae6f1966a4d91d362669fafa3d95526b132a6341e3dfff6447e0e76a07b3a7cfa6e8034574266b382b8e5174477ab8a32a49a57eda74895578031cd2d41fd0aef446046d6e633f5eb68a93013dfac1420bf7a1e1bf7a87476024478e97a1cc115de9`
-var signPrivatekey = `18c09c2639d7c8b3f26b273cdbfddf330c4f86c2ac3030a6b9a8533dc0c91f5e`
-var data = `TEST`
-
-func TestSSSignAndVerify(t *testing.T) {
-	signScheme := zcncrypto.NewSignatureScheme("bls0chain")
-	signScheme.SetPrivateKey(signPrivatekey)
-	hash := zcncrypto.Sha3Sum256(data)
-
-	fmt.Println("hash", hash)
-	fmt.Println("privkey", signScheme.GetPrivateKey())
-
-	var sk bls.SecretKey
-	sk.DeserializeHexStr(signScheme.GetPrivateKey())
-	pk := sk.GetPublicKey()
-	fmt.Println("pubkey", pk.ToString())
-
-	signature, err := signScheme.Sign(hash)
-
-	fmt.Println("signature", signature)
-
-	if err != nil {
-		t.Fatalf("BLS signing failed")
-	}
-	verifyScheme := zcncrypto.NewSignatureScheme("bls0chain")
-	verifyScheme.SetPublicKey(verifyPublickey)
-	if ok, err := verifyScheme.Verify(signature, hash); err != nil || !ok {
-		t.Fatalf("Verification failed\n")
-	}
-}
 
 func validateClientDetails(allocation, clientJSON string) error {
 	if len(allocation) == 0 || len(clientJSON) == 0 {
@@ -301,7 +267,6 @@ func Download(this js.Value, p []js.Value) interface{} {
 		return nil
 	})
 
-	// Create and return the Promise object
 	promiseConstructor := js.Global().Get("Promise")
 	return promiseConstructor.New(handler)
 }
@@ -352,14 +317,12 @@ func Delete(this js.Value, p []js.Value) interface{} {
 			responseConstructor := js.Global().Get("Response")
 			response := responseConstructor.New(js.ValueOf("Delete done successfully"))
 
-			// Resolve the Promise
 			resolve.Invoke(response)
 		}()
 
 		return nil
 	})
 
-	// Create and return the Promise object
 	promiseConstructor := js.Global().Get("Promise")
 	return promiseConstructor.New(handler)
 }
@@ -489,14 +452,12 @@ func Rename(this js.Value, p []js.Value) interface{} {
 			responseConstructor := js.Global().Get("Response")
 			response := responseConstructor.New(js.ValueOf("Rename done successfully"))
 
-			// Resolve the Promise
 			resolve.Invoke(response)
 		}()
 
 		return nil
 	})
 
-	// Create and return the Promise object
 	promiseConstructor := js.Global().Get("Promise")
 	return promiseConstructor.New(handler)
 }
@@ -547,14 +508,12 @@ func Copy(this js.Value, p []js.Value) interface{} {
 			responseConstructor := js.Global().Get("Response")
 			response := responseConstructor.New(js.ValueOf("Copy done successfully"))
 
-			// Resolve the Promise
 			resolve.Invoke(response)
 		}()
 
 		return nil
 	})
 
-	// Create and return the Promise object
 	promiseConstructor := js.Global().Get("Promise")
 	return promiseConstructor.New(handler)
 }
@@ -629,17 +588,12 @@ func Share(this js.Value, p []js.Value) interface{} {
 			responseConstructor := js.Global().Get("Response")
 			response := responseConstructor.New(js.ValueOf(at))
 
-			// var response struct {
-			// 	AuthTicket string `json:"auth_ticket"`
-			// }
-			// response.AuthTicket = at
 			resolve.Invoke(response)
 		}()
 
 		return nil
 	})
 
-	// Create and return the Promise object
 	promiseConstructor := js.Global().Get("Promise")
 	return promiseConstructor.New(handler)
 }
@@ -690,16 +644,12 @@ func Move(this js.Value, p []js.Value) interface{} {
 			responseConstructor := js.Global().Get("Response")
 			response := responseConstructor.New(js.ValueOf("Move done successfully"))
 
-			// resp := response{
-			// 	Message: "Move done successfully",
-			// }
 			resolve.Invoke(response)
 		}()
 
 		return nil
 	})
 
-	// Create and return the Promise object
 	promiseConstructor := js.Global().Get("Promise")
 	return promiseConstructor.New(handler)
 }
@@ -802,16 +752,12 @@ func Upload(this js.Value, p []js.Value) interface{} {
 			responseConstructor := js.Global().Get("Response")
 			response := responseConstructor.New(js.ValueOf("Upload done successfully"))
 
-			// resp := response{
-			// 	Message: "Upload done successfully",
-			// }
 			resolve.Invoke(response)
 		}()
 
 		return nil
 	})
 
-	// Create and return the Promise object
 	promiseConstructor := js.Global().Get("Promise")
 	return promiseConstructor.New(handler)
 }
@@ -819,9 +765,6 @@ func Upload(this js.Value, p []js.Value) interface{} {
 //-----------------------------------------------------------------------------
 
 func main() {
-	// Ported over a basic unit test to make sure it runs in the browser.
-	// TestSSSignAndVerify(new(testing.T))
-
 	fmt.Printf("0CHAIN - GOSDK (version=%v)\n", version.VERSIONSTR)
 
 	c := make(chan struct{}, 0)
@@ -844,7 +787,6 @@ func main() {
 	js.Global().Set("ConvertZcnTokenToETH", js.FuncOf(ConvertZcnTokenToETH))
 	js.Global().Set("SuggestEthGasPrice", js.FuncOf(SuggestEthGasPrice))
 	js.Global().Set("TransferEthTokens", js.FuncOf(TransferEthTokens))
-
 	js.Global().Set("GetWalletAddrFromEthMnemonic", js.FuncOf(GetWalletAddrFromEthMnemonic))
 	js.Global().Set("IsValidEthAddress", js.FuncOf(IsValidEthAddress))
 	js.Global().Set("CreateWalletFromEthMnemonic", js.FuncOf(CreateWalletFromEthMnemonic))
@@ -889,16 +831,16 @@ func main() {
 	js.Global().Set("GetMinerSCNodePool", js.FuncOf(GetMinerSCNodePool))
 	js.Global().Set("GetMinerSCUserInfo", js.FuncOf(GetMinerSCUserInfo))
 	js.Global().Set("GetMinerSCConfig", js.FuncOf(GetMinerSCConfig))
-	js.Global().Set("GetStorageSCConfig", js.FuncOf(GetStorageSCConfig))
-	js.Global().Set("GetChallengePoolInfo", js.FuncOf(GetChallengePoolInfo))
-	js.Global().Set("GetAllocation", js.FuncOf(GetAllocation))
-	js.Global().Set("GetAllocations", js.FuncOf(GetAllocations))
-	js.Global().Set("GetReadPoolInfo", js.FuncOf(GetReadPoolInfo))
-	js.Global().Set("GetStakePoolInfo", js.FuncOf(GetStakePoolInfo))
-	js.Global().Set("GetStakePoolUserInfo", js.FuncOf(GetStakePoolUserInfo))
-	js.Global().Set("GetBlobbers", js.FuncOf(GetBlobbers))
-	js.Global().Set("GetBlobber", js.FuncOf(GetBlobber))
-	js.Global().Set("GetWritePoolInfo", js.FuncOf(GetWritePoolInfo))
+	js.Global().Set("GetWalletStorageSCConfig", js.FuncOf(GetWalletStorageSCConfig))
+	js.Global().Set("GetWalletChallengePoolInfo", js.FuncOf(GetWalletChallengePoolInfo))
+	js.Global().Set("GetWalletAllocation", js.FuncOf(GetWalletAllocation))
+	js.Global().Set("GetWalletAllocations", js.FuncOf(GetWalletAllocations))
+	js.Global().Set("GetWalletReadPoolInfo", js.FuncOf(GetWalletReadPoolInfo))
+	js.Global().Set("GetWalletStakePoolInfo", js.FuncOf(GetWalletStakePoolInfo))
+	js.Global().Set("GetWalletStakePoolUserInfo", js.FuncOf(GetWalletStakePoolUserInfo))
+	js.Global().Set("GetWalletBlobbers", js.FuncOf(GetWalletBlobbers))
+	js.Global().Set("GetWalletBlobber", js.FuncOf(GetWalletBlobber))
+	js.Global().Set("GetWalletWritePoolInfo", js.FuncOf(GetWalletWritePoolInfo))
 	js.Global().Set("Encrypt", js.FuncOf(Encrypt))
 	js.Global().Set("Decrypt", js.FuncOf(Decrypt))
 
@@ -906,35 +848,33 @@ func main() {
 	js.Global().Set("initializeConfig", js.FuncOf(InitializeConfig))
 	js.Global().Set("initStorageSDK", js.FuncOf(InitStorageSDK))
 	js.Global().Set("InitAuthTicket", js.FuncOf(InitAuthTicket))
-	js.Global().Set("SetSDKLogLevel", js.FuncOf(SetSDKLogLevel))
-	js.Global().Set("SetSDKLogFile", js.FuncOf(SetSDKLogFile))
 	js.Global().Set("GetNetwork", js.FuncOf(GetNetwork))
 	js.Global().Set("SetMaxTxnQuery", js.FuncOf(SetMaxTxnQuery))
 	js.Global().Set("SetQuerySleepTime", js.FuncOf(SetQuerySleepTime))
 	js.Global().Set("SetMinSubmit", js.FuncOf(SetMinSubmit))
 	js.Global().Set("SetMinConfirmation", js.FuncOf(SetMinConfirmation))
-	js.Global().Set("ZBOXSetNetwork", js.FuncOf(ZBOXSetNetwork))
+	js.Global().Set("SetNetwork", js.FuncOf(SetNetwork))
 	js.Global().Set("CreateReadPool", js.FuncOf(CreateReadPool))
 	js.Global().Set("AllocFilter", js.FuncOf(AllocFilter))
-	js.Global().Set("ZBOXGetReadPoolInfo", js.FuncOf(ZBOXGetReadPoolInfo))
+	js.Global().Set("GetReadPoolInfo", js.FuncOf(GetReadPoolInfo))
 	js.Global().Set("ReadPoolLock", js.FuncOf(ReadPoolLock))
 	js.Global().Set("ReadPoolUnlock", js.FuncOf(ReadPoolUnlock))
-	js.Global().Set("ZBOXGetStakePoolInfo", js.FuncOf(ZBOXGetStakePoolInfo))
-	js.Global().Set("ZBOXGetStakePoolUserInfo", js.FuncOf(ZBOXGetStakePoolUserInfo))
+	js.Global().Set("GetStakePoolInfo", js.FuncOf(GetStakePoolInfo))
+	js.Global().Set("GetStakePoolUserInfo", js.FuncOf(GetStakePoolUserInfo))
 	js.Global().Set("StakePoolLock", js.FuncOf(StakePoolLock))
 	js.Global().Set("StakePoolUnlock", js.FuncOf(StakePoolUnlock))
 	js.Global().Set("StakePoolPayInterests", js.FuncOf(StakePoolPayInterests))
-	js.Global().Set("ZBOXGetWritePoolInfo", js.FuncOf(ZBOXGetWritePoolInfo))
+	js.Global().Set("GetWritePoolInfo", js.FuncOf(GetWritePoolInfo))
 	js.Global().Set("WritePoolLock", js.FuncOf(WritePoolLock))
 	js.Global().Set("WritePoolUnlock", js.FuncOf(WritePoolUnlock))
-	js.Global().Set("ZBOXGetChallengePoolInfo", js.FuncOf(ZBOXGetChallengePoolInfo))
-	js.Global().Set("ZBOXGetStorageSCConfig", js.FuncOf(ZBOXGetStorageSCConfig))
-	js.Global().Set("ZBOXGetBlobbers", js.FuncOf(ZBOXGetBlobbers))
-	js.Global().Set("ZBOXGetBlobber", js.FuncOf(ZBOXGetBlobber))
+	js.Global().Set("GetChallengePoolInfo", js.FuncOf(GetChallengePoolInfo))
+	js.Global().Set("GetStorageSCConfig", js.FuncOf(GetStorageSCConfig))
+	js.Global().Set("GetBlobbers", js.FuncOf(GetBlobbers))
+	js.Global().Set("GetBlobber", js.FuncOf(GetBlobber))
 	js.Global().Set("GetClientEncryptedPublicKey", js.FuncOf(GetClientEncryptedPublicKey))
 	js.Global().Set("GetAllocationFromAuthTicket", js.FuncOf(GetAllocationFromAuthTicket))
-	js.Global().Set("ZBOXGetAllocation", js.FuncOf(ZBOXGetAllocation))
-	js.Global().Set("ZBOXGetAllocations", js.FuncOf(ZBOXGetAllocations))
+	js.Global().Set("GetAllocation", js.FuncOf(GetAllocation))
+	js.Global().Set("GetAllocations", js.FuncOf(GetAllocations))
 	js.Global().Set("GetAllocationsForClient", js.FuncOf(GetAllocationsForClient))
 	js.Global().Set("CreateAllocation", js.FuncOf(CreateAllocation))
 	js.Global().Set("CreateAllocationForOwner", js.FuncOf(CreateAllocationForOwner))
